ls-tree: add --sparse-filter-oid argument
There is currently no straight-forward way to ask git "which files would
be part of a sparse checkout of commit X with sparse checkout patterns
Y". A query like this can be useful for tooling wanting to discern
whether sparse checkouts two commits contain the same content even if
the full trees differ.

'rev-list --objects --filter=sparse:oid' comes close, but as rev-list is
concerned with objects rather than directory trees, it leaves files when
the same blob occurs in at two different paths.

Similarly, ls-files also have some of the functionality. However, it gets
awkward when you are interested in another commit or sparsity
patterns that aren't currently in '.git/info/sparse-checkout'.

To fill this gap, teach ls-tree to be able to filter its output
based on sparsity patterns. The added argument --sparse-filter-oid takes
the object id of a blob containing a sparsity patterns, which
then are used to filter which paths are output.

When provided sparsity parameters, the output ls-tree will only output
blobs. While it may be valid in some situations to output a tree object
-- e.g. when a cone pattern matches all blobs recursively contained in a
tree -- it is less unclear what should be output if a sparse pattern
matches parts of a tree. Thus ls-tree with sparsity patterns is mostly
useful with '-r', but this implementation also leaves the option of
querying for "which files in a given tree object would be part of a
sparse checkout".

To allow for reusing the pattern matching logic found in
'path_in_sparse_checkout_1' from 'dir.c' with arbitrary patterns, the
pattern matching part of that function is extracted into a new function
'recursively_match_path_with_sparse_patterns'.